+++ src/test/java/fr/marcwrobel/jbanking/creditor/CreditorIdentifierTest.java
@@ -217,7 +217,7 @@ class CreditorIdentifierTest {
     String nationalId = value.substring(7);
     assertEquals(
         creditorId,
-        new CreditorIdentifier(IsoCountry.fromCode(countryCode), businessCode, nationalId));
+        new CreditorIdentifier(IsoCountry.fromAlpha2Code(countryCode), businessCode, nationalId));
   }
 
   @ParameterizedTest
@@ -242,7 +242,7 @@ class CreditorIdentifierTest {
     assertEquals(
         value,
         new CreditorIdentifier(
-                IsoCountry.fromCode(countryCode), businessCode, nationalId.toLowerCase())
+                IsoCountry.fromAlpha2Code(countryCode), businessCode, nationalId.toLowerCase())
             .toString());
   }